Move definitions for template SLB_level_conv.bt to a single file
